requirements.txt
coverage>=2.85
flake8>=3
mock
nose>=0.11.1
nosexcover>=1.0.7
python-dateutil
six>=1.9.0
